pyproject.toml
[project]
name = "langgraph-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dashscope>=1.24.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "openai>=1.97.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tavily-python>=0.7.10",
]
